StellationController running as Javascript
I did major surgery to factor StellationController out of StellationMain.  The boundary
may not be perfect, certainly not limited to serializable types as a worker would need.
Everything compiles for Java, and seems to run OK.

I changed FastHashtable to extend Hashtable rather than Dictionary, since the latter
was turned into an interface by JSweet.

JSweet is past the javac errors, now getting 98 Typescript errors.  I'm ignoring them
for the moment, and testing the runtime.

There is no StreamTokenizer implementation in j4ts, so I used the
FixedStreamTokenizer that is in this source.  I had to work around some
JSweet overloaded-constructor defect, but it seems to be working both in Java
and in Javascript... I am able to load the u27.off file.

I renamed all the off-as-gif resources, since the old Netscape defect is moot.

I had to shim java.io.PrintStream.printf, and I did it badly, but the data comes out
on the console.